requirements.txt
aiohttp==3.6.2
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
discord.py==1.3.3
idna==2.9
multidict==4.7.5
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websockets==8.1
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.4.2
